boot: refactor i2c driver into namespace
