+++ stratosphere/boot/source/boot_pcv.cpp
@@ -1,0 +1,77 @@
+/*
+ * Copyright (c) 2018-2019 Atmosphère-NX
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms and conditions of the GNU General Public License,
+ * version 2, as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include <switch.h>
+#include <stratosphere.hpp>
+
+#include "i2c/i2c_types.hpp"
+#include "i2c/driver/impl/i2c_pcv.hpp"
+#include "i2c/driver/impl/i2c_registers.hpp"
+
+using namespace sts::i2c::driver::impl;
+
+namespace sts::pcv {
+
+    void Initialize() {
+        /* Don't do anything. */
+    }
+
+    void Finalize() {
+        /* Don't do anything. */
+    }
+
+    Result SetClockRate(PcvModule module, u32 hz) {
+        /* Get clock/reset registers. */
+        ClkRstRegisters regs;
+        regs.SetBus(ConvertFromPcvModule(module));
+        /* Set clock enabled/source. */
+        SetRegisterBits(regs.clk_en_reg, regs.mask);
+        ReadWriteRegisterBits(regs.clk_src_reg, 0x4, 0xFF);
+        svcSleepThread(1000ul);
+        ReadWriteRegisterBits(regs.clk_src_reg, 0, 0xE0000000);
+        svcSleepThread(2000ul);
+
+        return ResultSuccess;
+    }
+
+    Result SetClockEnabled(PcvModule module, bool enabled) {
+        return ResultSuccess;
+    }
+
+    Result SetVoltageEnabled(u32 domain, bool enabled) {
+        return ResultSuccess;
+    }
+
+    Result SetVoltageValue(u32 domain, u32 voltage) {
+        return ResultSuccess;
+    }
+
+    Result SetReset(PcvModule module, bool reset) {
+        /* Get clock/reset registers. */
+        ClkRstRegisters regs;
+        regs.SetBus(ConvertFromPcvModule(module));
+
+        /* Set/clear reset. */
+        if (reset) {
+            SetRegisterBits(regs.rst_reg, regs.mask);
+        } else {
+            ClearRegisterBits(regs.rst_reg, ~regs.mask);
+        }
+
+        return ResultSuccess;
+    }
+
+}
